testes de interação da funcionalidade soma
