+++ calculadora/api/src/test/java/com/danielfireman/ifal/calcapi/IntegrationTest.java
@@ -29,29 +29,6 @@ public class IntegrationTest {
   }
 
   @Test
-  public void checaErro(int serverPort) throws IOException {
-    Request req = new Request.Builder()
-        .url("http://localhost:" + serverPort + "/soma/a/a")
-        .build();
-
-    try (Response rsp = client.newCall(req).execute()) {
-      assertEquals(StatusCode.BAD_REQUEST_CODE, rsp.code());
-    }
-  }
-
-  @Test
-  public void checaSucesso(int serverPort) throws IOException {
-    Request req = new Request.Builder()
-        .url("http://localhost:" + serverPort + "/soma/10/10")
-        .build();
-
-    try (Response rsp = client.newCall(req).execute()) {
-      assertEquals(StatusCode.OK_CODE, rsp.code());
-      assertEquals("20.0", rsp.body().string());
-    }
-  }
-
-  @Test
   public void checError (int serverPort) throws IOException {
     Request req = new Request.Builder()
         .url("http://localhost:" + serverPort + "/div/a/a")
